refactor(AudioAsset): create AudioAssetRestricted component and update imports
