+++ src/script/components/MessagesList/Message/ContentMessage/asset/AudioAsset/AudioAssetRestricted/AudioAssetRestricted.styles.ts
