+++ src/script/components/MessagesList/Message/ContentMessage/asset/AudioAsset/AudioAssetSeekBar/AudioSeekBarNew/useAudioSeekBar/useAudioSeekBar.ts
@@ -118,10 +118,6 @@ export const useAudioSeekBar = ({asset, audioElement, svgRef}: UseAudioSeekBarPr
   }, [asset]);
 
   useEffect(() => {
-    if (!audioElement) {
-      return;
-    }
-
     audioElement.addEventListener('ended', onAudioEnded);
     audioElement.addEventListener('timeupdate', onTimeUpdate);
